fix: empId & sql

+++ database/v2020_07_06_history_log.sql
@@ -4,7 +4,7 @@ CREATE TABLE `history_log` (
   `entityId` int NOT NULL,
   `changeLog` text COLLATE utf8mb4_unicode_ci NOT NULL,
   `relatedJSONData` text COLLATE utf8mb4_unicode_ci,
-  `userId` int NOT NULL,
+  `userId` bigint(11) unsigned NOT NULL,
   `createdAt` datetime NOT NULL,
   `updatedAt` datetime NOT NULL,
   `deletedAt` datetime DEFAULT NULL,
